requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
gevent==24.11.1
greenlet==3.1.1
idna==3.10
pycparser==2.22
pytz==2025.1
requests==2.32.3
setuptools==75.8.0
urllib3==2.3.0
websocket-client==1.8.0
zope.event==5.0
zope.interface==7.2
